@ #1 Increased the highlightPriority to 0
Since the plugin was clashing with CocActionAsync('highlight'). I've
inreased the priority of this plugin's highlighting. from -1 to 0. It
was set to -1 before so that it didn't clash with the built in vim
search highlighting, which the documentation says is set to 0. However
after settign the highlightPriority to 0. It does seem to work with the
highlithting just fine (at least for my current versions of vim and nvim).
I'll leave it set to 0 for the time being.

I've also made it possible for the user to configure highlightPriority
from there .vimrc/init.vim file. This can be done by simpling doing the
following:
    let g:highlightPriority=-1