See #3 fixed for you-n-g
Unfornetly you-n-g and my vims seem to be behaving differently.
When using '< and '> to get the visual selection you-n-g's works without
alteration. On my setup I need to -1 off the end column to get this to
work. Something feels wrong with mine. I'll change it to the way that
works your you-n-g. However I will need to find a solution that works
for both. I'd like to not need to set a specal option in my vimrc to get
it working in my setup.