See #5, fix when vim's ignorecase  is on
Fixed a string comparison so it ignores the vim session's "ignorecase"
option and compares the strings using a case sesntivie comaprision.

Without this fix the plugin through that asdf and ASDF were both the
same and so one would overide the other. Since even though the matching
was not case sensitive the highlighting was.